Improve bro_bif_create_loader function.
It now generates the __load__.bro script from an explicit list of
scripts that are passed to the function as an argument instead of trying
to generate it from whatever files are in the directory at build-time
(which may become stale).

The bif_target macro also now ends up generating two such lists (which
should always be up-to-date) that can be used as arguments for
bro_bif_create_loader.