Add CMake script to check if compiler can do C++11
I don't like this much.  Enumerating all compilers explicitly isn't
great, but I guess it can help a little for the common compilers.  But
also the compiler version detection is dependent on CMake version, which
may make it less reliable/consistent.  And when CMake doesn't provide
it, falling back to manually detecting the version from command line
output is also fragile since some compilers may not provide a straight
forward way to do that (e.g. for Clang, best way I can think is to rely
on a extra program which emits the version string as constructed via
preprocessor definitions).