Add option to find only static CAF libraries
If the global CMake variable CAF_STATIC_ONLY is set, FindCAF.cmake
attempts to locate only the static libraries of CAF.

This patch also fixes an issue when CAF_ROOT_DIR was specified but CAF
has already been installed in system installation path. To do so,
find_path needs to called twice, because the HINTS directive only
applies when CMake can't find anything in the system path. This is not
pretty, but there's no other solution to the best of our knowledge.

Patch provided by Steffen Haas.