Fix some cmake+OpenSSL problems.
On OS-X, --with-openssl failed to override the OpenSSL installation in
macports/homebrew. The reason for this is that CMAKE_PREFIX_PATH, which
is set on OS-X, has a higher priority than the hint that is passed to
find_*.

The only working solution that I found is to unset CMAKE_PREFIX_PATH for
find_package in case the package root dir is explicitly set.

This patch also aborts configure if OpenSSL 1.1 is found - which we do
not compile with yet. And it rewrites the old OpenSSL version test to
just use the determined version.