GH-5: add libl dependency for OpenSSL tests
This generally should probably be built in to CMake's FindOpenSSL
module such that it includes it within OPENSSL_LIBRARIES automatically
in the necessary cases (e.g. when using static OpenSSL libs).  A CMake
patch to do this may have been suggested before, but not sure what the
status of it is:

https://cmake.org/pipermail/cmake-developers/2016-August/029210.html