Make changes to CMAKE_CXX_FLAGS globally visible
When setting variables in an included file, CMake uses "dynamic
scoping". This seems to cause the changes to remain visible in the
including scope (Zeek's `CMakeLists.txt`) but does not extend to
subdirectories. Changing the CMAKE_CXX_FLAGS property explicitly with
global context seems to fix this issue and allow the flags to propagate
to subdirectories, in particular to Broker.