BifCl.cmake: generate include files during configuration
Functions bro_bif_create_includes and bro_bif_create_register add
targets that create files during the build.  The targets produced by a
MinGW build of CMake with the Ninja generator fail because it uses
cmd.exe as the shell.  Address this (and simplify a bit) by creating the
files during configuration instead.