Set correct OUTPUTS for binary install package.
The custom command for binary install packages declares an `OUTPUT`
that it produced `${plugin_name_canon}.tar.gz` in the current bin dir.
This was not correct as `zeek-plugin-create-package.sh` produces a
`tar.gz` file in the `dist` subdirectory which it symlinks into the
current bin dir as `tgz`. Due to that the output of the command was
never present and the `dist` was always run.

This patch fixes this by changing the custom command to declare that it
creates the file in the current bin directory (the `tgz` file). With
that the rule producing the file is only rerun if the inputs of the
custom command become outdated.