Zeek/Zeek Issue #1848
Add new function zeek_plugin_script().

Script files added through this function will be added as dependencies
to the binary package, and changes to the script will be picked up
by `make && sudo make install` without a `make clean` being required.

Scripts that are *not* added through this command will still be
distributed as before, but a change to the script will require
a `make clean` ( as before ) in order to trigger repackaging if
there is no change that triggers the plugin's not script components
to build.

Follow ons to this are updating the zeek reference, adding init-plugin
support and any doc changes