Fix link handling with binary packaging in InstallSymlink
With binary packaging, this moves the symlink creation from configuration to
installation, uses an alternative name for the symlink created in the build
directory prior to installation, and removes this link after installation to
avoid dangling links.

This fixes two problems: (1) the early symlink creation could interfere with
supposedly local operations on the file, for example creating a symlink to
itself, (2) leaving behind symlinks into the build tree after one install could
trip up subsequent configure/build runs because again operations intended to be
local ended up in the build tree, which might not be writable or not exist at
all.