Prevent build interference from creating dist tarball.
We previously would copy any additional files into the build directory.
Since the build directory is e.g., in the C++ include paths this caused
issues if e.g., a named of an additional file matched a header (concrete
issue: `VERSION` file interferes with `<version>` header which exists
for C++20-capable compilers).

With this patch we instead assemble files for distribution in a
dedicated staging directory. This also allows us to simplify the code
since we do not need to deal with `MANIFEST` files anymore.

Closes zeek/zeek#2784.