BinPAC: Ensure zeek/src/include is added to BinPAC include path
We've removed the zeek -> . symlink from zeek/src and now create
it dynamically during the build in zeek/src/include. Ensure that
new directory is in BinPAC's include path, too.

Fixes --include-plugins with plugins that contain BinPAC based analyzers
and using `%include zeek/binpac.pac` as, for example, this one:

     https://github.com/cisagov/icsnpp-bacnet.git