Prefer to find `spicy-config` in configured over default search paths.
When searching for `spicy-config` we previously we would pass configured
paths as `HINTS` to `find_program`. Since `HINTS` adds additional search
paths to builtin search paths, this did not find the correct
`spicy-config` if Zeek was configured with `--with-spicy=<SPICY_PREFIX>`
and another `spicy-config` was found in builtin paths; instead we would
always prefer to use the one from builtin paths.

With this patch we split the search for `spicy-config` into two phases,
one using exclusively the configured paths and only if that search did
not succeed a second phase which would look in builtin paths. With that
we prefer a `spicy-config` from an explicitly configured path.