ZeekPlugin: Make non-sufficient minimum version a FATAL_ERROR
The warning is useful, but usually scrolls off the screen before users
run into errors that are less actionable. We now had 3 reports of users
that were stuck with less useful message around the IN_LIST operator
when all they had to do was but the cmake_minimum_required() line.