Do not leak dependencies of plugins
Previously plugin dependencies declared with `DEPENDENCIES` would be
`PUBLIC` to the plugin's target. This means that e.g., include
directories from the plugin's dependency would be added to any of its
dependents as well.

This not only increases the surface of what we might pull in
inadvertently elsewhere, but also might cause unneeded rebuilds
elsewhere if dependency paths change.

With this patch we declare any such dependencies `PRIVATE`. This should
only break code which previously did not correctly declare all its
dependencies. This code is used both for internal as well is external
plugins, and it seems to work well for internal code. For user code this
is technically a breaking change, but I expect complex CMake
dependencies between multiple plugins to be rare.